Moved ITimerManager from Timers to Timing.
