+++ Robust.Shared/SharedIoC.cs
@@ -13,7 +13,6 @@ using Robust.Shared.Physics;
 using Robust.Shared.Random;
 using Robust.Shared.Sandboxing;
 using Robust.Shared.Serialization;
-using Robust.Shared.Timers;
 using Robust.Shared.Timing;
 
 namespace Robust.Shared
